put CircleAndImages into folder, remove CircleTitleImages, rename useState to appState
