+++ components/CircleAndImages/PinballImage.tsx
@@ -1,19 +1,19 @@
-import ImageModel from "../models/images";
+import ImageModel from "../../models/images";
 
 const PinballImage: React.FC<ImageModel> = ({ onHelloPage }) => (
   <>
     <picture>
       <source
-        srcSet={require("../public/images/pinball.png?webp")}
+        srcSet={require("../../public/images/pinball.png?webp")}
         type="image/webp"
       />
       <source
-        srcSet={require("../public/images/pinball.png")}
+        srcSet={require("../../public/images/pinball.png")}
         type="image/png"
       />
       <img
         className={`pinball ${onHelloPage ? "blueFilter" : "redFilter"}`}
-        srcSet={require("../public/images/pinball.png")}
+        srcSet={require("../../public/images/pinball.png")}
         alt="A pinball table"
       />
     </picture>
